Address Codacy comments.
