+++ Core/src/org/sleuthkit/autopsy/datamodel/ImageNode.java
@@ -28,12 +28,15 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.logging.Level;
 import javax.swing.Action;
+import org.apache.commons.lang3.tuple.Pair;
 import org.openide.nodes.Sheet;
 import org.openide.util.NbBundle;
 import org.openide.util.NbBundle.Messages;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.casemodule.NoCurrentCaseException;
 import org.sleuthkit.autopsy.casemodule.datasourcesummary.ViewSummaryInformationAction;
+import org.sleuthkit.autopsy.centralrepository.datamodel.CorrelationAttributeInstance;
+import org.sleuthkit.autopsy.corecomponents.DataResultViewerTable;
 import org.sleuthkit.autopsy.coreutils.Logger;
 import org.sleuthkit.autopsy.directorytree.ExplorerNodeActionVisitor;
 import org.sleuthkit.autopsy.directorytree.FileSearchAction;
@@ -47,6 +50,7 @@ import org.sleuthkit.datamodel.SleuthkitCase.CaseDbQuery;
 import org.sleuthkit.datamodel.TskCoreException;
 import org.sleuthkit.datamodel.VirtualDirectory;
 import org.sleuthkit.autopsy.datamodel.BaseChildFactory.NoSuchEventBusException;
+import org.sleuthkit.datamodel.Tag;
 
 /**
  * This class is used to represent the "Node" for the image. The children of
@@ -250,4 +254,67 @@ public class ImageNode extends AbstractContentNode<Image> {
         }
     };
 
+    /**
+     * Reads and returns a list of all tags associated with this content node.
+     *
+     * Null implementation of an abstract method. 
+     * 
+     * @return list of tags associated with the node.
+     */ 
+    @Override
+    protected List<Tag> getAllTagsFromDatabase() {
+        return new ArrayList<>();
+    }
+    /**
+     * Returns correlation attribute instance for the underlying content of the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @return correlation attribute instance for the underlying content of the node.
+     */ 
+    @Override
+    protected CorrelationAttributeInstance getCorrelationAttributeInstance() {
+        return null;
+    }
+    
+    /**
+     * Returns Score property for the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @param tags list of tags.
+     * 
+     * @return Score property for the underlying content of the node.
+     */ 
+    @Override
+    protected Pair<DataResultViewerTable.Score, String> getScorePropertyAndDescription(List<Tag> tags) {
+        return Pair.of(DataResultViewerTable.Score.NO_SCORE, NO_DESCR);
+    }
+    /**
+     * Returns comment property for the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @param tags list of tags
+     * @param attribute correlation attribute instance
+     * 
+     * @return Comment property for the underlying content of the node.
+     */ 
+    @Override
+    protected DataResultViewerTable.HasCommentStatus getCommentProperty(List<Tag> tags, CorrelationAttributeInstance attribute) {
+        return DataResultViewerTable.HasCommentStatus.NO_COMMENT;
+    }
+    /**
+     * Returns occurrences/count property for the node.
+     *
+     * Null implementation of an abstract method. 
+     * 
+     * @param attribute correlation attribute instance
+     * 
+     * @return count property for the underlying content of the node.
+     */  
+    @Override
+    protected Pair<Long, String> getCountPropertyAndDescription(CorrelationAttributeInstance attribute) {
+        return Pair.of(-1L, NO_DESCR);
+    }
 }
