+++ Core/src/org/sleuthkit/autopsy/datamodel/VolumeNode.java
@@ -25,10 +25,14 @@ import java.util.EnumSet;
 import java.util.List;
 import java.util.logging.Level;
 import javax.swing.Action;
+import org.apache.commons.lang3.tuple.Pair;
 import org.openide.nodes.Sheet;
 import org.openide.util.NbBundle;
 import org.sleuthkit.autopsy.casemodule.Case;
+import org.sleuthkit.autopsy.centralrepository.datamodel.CorrelationAttributeInstance;
+import org.sleuthkit.autopsy.corecomponents.DataResultViewerTable;
 import org.sleuthkit.autopsy.coreutils.Logger;
+import static org.sleuthkit.autopsy.datamodel.AbstractContentNode.NO_DESCR;
 import org.sleuthkit.autopsy.datamodel.BaseChildFactory.NoSuchEventBusException;
 import org.sleuthkit.autopsy.directorytree.ExplorerNodeActionVisitor;
 import org.sleuthkit.autopsy.directorytree.NewWindowViewAction;
@@ -39,6 +43,7 @@ import org.sleuthkit.datamodel.TskCoreException;
 import org.sleuthkit.datamodel.VirtualDirectory;
 import org.sleuthkit.datamodel.Volume;
 import org.sleuthkit.autopsy.directorytree.FileSystemDetailsAction;
+import org.sleuthkit.datamodel.Tag;
 
 /**
  * This class is used to represent the "Node" for the volume. Its child is the
@@ -211,5 +216,68 @@ public class VolumeNode extends AbstractContentNode<Volume> {
     @Override
     public String getItemType() {
         return DisplayableItemNode.FILE_PARENT_NODE_KEY;
+    }
+    /**
+     * Reads and returns a list of all tags associated with this content node.
+     *
+     * Null implementation of an abstract method. 
+     * 
+     * @return list of tags associated with the node.
+     */ 
+    @Override
+    protected List<Tag> getAllTagsFromDatabase() {
+        return new ArrayList<>();
+    }
+    /**
+     * Returns correlation attribute instance for the underlying content of the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @return correlation attribute instance for the underlying content of the node.
+     */ 
+    @Override
+    protected CorrelationAttributeInstance getCorrelationAttributeInstance() {
+        return null;
+    }
+    
+    /**
+     * Returns Score property for the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @param tags list of tags.
+     * 
+     * @return Score property for the underlying content of the node.
+     */ 
+    @Override
+    protected Pair<DataResultViewerTable.Score, String> getScorePropertyAndDescription(List<Tag> tags) {
+        return Pair.of(DataResultViewerTable.Score.NO_SCORE, NO_DESCR);
+    }
+    /**
+     * Returns comment property for the node.
+     * 
+     * Null implementation of an abstract method. 
+     * 
+     * @param tags list of tags
+     * @param attribute correlation attribute instance
+     * 
+     * @return Comment property for the underlying content of the node.
+     */ 
+    @Override
+    protected DataResultViewerTable.HasCommentStatus getCommentProperty(List<Tag> tags, CorrelationAttributeInstance attribute) {
+        return DataResultViewerTable.HasCommentStatus.NO_COMMENT;
+    }
+    /**
+     * Returns occurrences/count property for the node.
+     *
+     * Null implementation of an abstract method. 
+     * 
+     * @param attribute correlation attribute instance
+     * 
+     * @return count property for the underlying content of the node.
+     */  
+    @Override
+    protected Pair<Long, String> getCountPropertyAndDescription(CorrelationAttributeInstance attribute) {
+        return Pair.of(-1L, NO_DESCR);
     }
 }
